게시글 목록 좋아요 수 추가

+++ src/main/java/com/example/ballkkaye/board/like/BoardLikeRepository.java
@@ -27,10 +27,10 @@ public class BoardLikeRepository {
         }
     }
 
-    public Long totalCount(Integer boardId) {
+    public Integer totalCount(Integer boardId) {
         return em.createQuery(
                         "SELECT COUNT(brl) FROM BoardLike brl WHERE brl.board.id = :boardId", Long.class)
                 .setParameter("boardId", boardId)
-                .getSingleResult();
+                .getSingleResult().intValue();
     }
 }
